1.3컨베이어 작업 Trap 박스 추가

+++ Assets/QuickSheet/Test/Runtime/makegoalData.cs
@@ -17,15 +17,15 @@ public class makegoalData
   public int[] Color { get {return color; } set { color = value;} }
   
   [SerializeField]
-  int[] x = new int[0];
-  public int[] X { get {return x; } set { x = value;} }
+  float[] x = new float[0];
+  public float[] X { get {return x; } set { x = value;} }
   
   [SerializeField]
-  int[] y = new int[0];
-  public int[] Y { get {return y; } set { y = value;} }
+  float[] y = new float[0];
+  public float[] Y { get {return y; } set { y = value;} }
   
   [SerializeField]
-  int[] z = new int[0];
-  public int[] Z { get {return z; } set { z = value;} }
+  float[] z = new float[0];
+  public float[] Z { get {return z; } set { z = value;} }
   
 }
