+++ Assets/QuickSheet/Test/Runtime/makestartData.cs
@@ -13,24 +13,24 @@ public class makestartData
   public int[] Num { get {return num; } set { num = value;} }
   
   [SerializeField]
-  int[] x = new int[0];
-  public int[] X { get {return x; } set { x = value;} }
+  float[] x = new float[0];
+  public float[] X { get {return x; } set { x = value;} }
   
   [SerializeField]
-  int[] y = new int[0];
-  public int[] Y { get {return y; } set { y = value;} }
+  float[] y = new float[0];
+  public float[] Y { get {return y; } set { y = value;} }
   
   [SerializeField]
-  int[] z = new int[0];
-  public int[] Z { get {return z; } set { z = value;} }
+  float[] z = new float[0];
+  public float[] Z { get {return z; } set { z = value;} }
   
   [SerializeField]
   int[] color = new int[0];
   public int[] Color { get {return color; } set { color = value;} }
   
   [SerializeField]
-  int[] time = new int[0];
-  public int[] Time { get {return time; } set { time = value;} }
+  float[] time = new float[0];
+  public float[] Time { get {return time; } set { time = value;} }
   
   [SerializeField]
   int[] count = new int[0];
